+++ CHANGES.rst
@@ -5,7 +5,7 @@ IP Navigator CHANGES
 development
 ===========
 - [mw] Fix Umlaut issues with DEPATISnet
-
+- Exclude .DS_Store files from the release tarball
 
 2018-09-26 0.151.0
 ==================