+++ CHANGES.rst
@@ -9,6 +9,7 @@ development
 - [mw] Strip DEPATISnet query expression before propagating to upstream. Thanks, Martin!
 - [mw] Grok error messages from erroneous DEPATISnet queries
 - [ui] Add new universal "start_search" method for kicking off a blank-slate search action
+- [ui] Run boot-time search after activating basket model
 
 
 2018-09-26 0.151.0