+++ CHANGES.rst
@@ -10,6 +10,7 @@ development
 - [mw] Grok error messages from erroneous DEPATISnet queries
 - [ui] Add new universal "start_search" method for kicking off a blank-slate search action
 - [ui] Run boot-time search after activating basket model
+- [ui] Limit query expression syntax chooser to DEPATISnet only
 
 
 2018-09-26 0.151.0