+++ CHANGES.rst
@@ -4,6 +4,10 @@ IP Navigator CHANGES
 
 development
 ===========
+
+
+2018-09-28 0.152.0
+==================
 - Exclude .DS_Store files from the release tarball
 - [mw] Fix Umlaut issues with DEPATISnet. Thanks, Martin!
 - [mw] Strip DEPATISnet query expression before propagating to upstream. Thanks, Martin!